// src/store/auth/types.ts
export enum AuthAction {
  LOGIN_REQUEST = 'LOGIN_REQUEST',
  REGISTER_REQUEST = 'REGISTER_REQUEST',
  CONFIRM_REQUEST = 'CONFIRM_REQUEST',
  RESEND_CONFIRAMTION_CODE = 'RESEND_CONFIRAMTION_CODE',
  FORGOT_PASSWORD = 'FORGOT_PASSWORD',
  RESET_PASSWORD = 'RESET_PASSWORD',

  LOGOUT = 'LOGOUT',
}

export const AuthScope = 'auth';